filebeat_operator/requirements.txt
boto3
rich
pyyaml
python_dotenv